changed: title anim starts when page is full loaded
