+++ src/pages/Home/Home.tsx
@@ -1,4 +1,4 @@
-import TitleScreen from "./components/TitleScreen"
+import TitleScreen from "./components/TitleScreen/TitleScreen"
 import Section from "../../components/Section"
 import AboutMe from "./components/AboutMe"
 import Abilities from "./components/Ablities"
@@ -18,7 +18,6 @@ export default function Home() {
             <TitleScreen />
             
             <div className="px-3 sm:px-5 md:px-7 lg:px-8">
-            {/* <div className="[margin-inline:_clamp(.8em,4vw,2em)]"> */}
                 <Section title="About me">
                     <AboutMe />
                 </Section>
